requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
structlog==24.1.0
tenacity==8.5.0
pybreaker==1.0.2
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.18
supabase==2.7.4
# Force a safe Authlib version to address CVEs reported by Trivy
authlib>=1.6.5
scipy==1.16.3
numpy==1.26.0
pandas==2.3.3
redis==5.1.0
slowapi==0.1.9
pypdf2==3.0.1
pymupdf==1.24.0
openai==1.52.0
rq==2.3.3
rq-scheduler==0.13.1
pyxirr==0.10.0
python-magic==0.4.27
clamd==1.0.2
aiofiles==24.1.0
email-validator==2.2.0
pydantic[email]==2.9.2
itsdangerous==2.2.0
python-json-logger==2.0.7
requests==2.32.3
google-auth==2.30.0
google-api-python-client==2.139.0
alembic==1.13.2
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
black==24.3.0
bleach==6.1.0
types-bleach==6.1.0.20240331
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-mock==3.14.0
pycryptodome==3.20.0
reportlab==4.2.5
better-profanity==0.7.0
python-statemachine==2.3.5
# observability + email (added 2026-02-20)
sentry-sdk[fastapi]==2.19.2
boto3==1.35.0
